Better read-only support
Pull Request: https://github.com/ViktorStolbovoy/MediaCollection-PP/pull/11 (master)
